TSK-1682: Moved Svelte deps to resource plugin
