+++ plugins/recruit/src/index.ts
@@ -19,19 +19,7 @@ import type { Asset, IntlString, Plugin, Resource } from '@hcengineering/platfor
 import { plugin } from '@hcengineering/platform'
 import type { ProjectTypeDescriptor, TaskType } from '@hcengineering/task'
 import { AnyComponent, ResolvedLocation } from '@hcengineering/ui'
-import type {
-  Applicant,
-  ApplicantMatch,
-  Candidate,
-  Candidates,
-  Opinion,
-  Review,
-  ScriptTypedAttributeEditorMixin,
-  ScriptTypedAttributeFactoryMixin,
-  ScriptTypedPropertyEditorMixin,
-  Vacancy,
-  VacancyList
-} from './types'
+import type { Applicant, ApplicantMatch, Candidate, Candidates, Opinion, Review, Vacancy, VacancyList } from './types'
 
 export * from './types'
 
@@ -60,10 +48,7 @@ const recruit = plugin(recruitId, {
   },
   mixin: {
     Candidate: '' as Ref<Mixin<Candidate>>,
-    VacancyList: '' as Ref<Mixin<VacancyList>>,
-    ScriptTypedAttributeEditor: '' as Ref<Mixin<ScriptTypedAttributeEditorMixin<any>>>,
-    ScriptTypedAttributeFactory: '' as Ref<Mixin<ScriptTypedAttributeFactoryMixin<any>>>,
-    ScriptTypedPropertyEditor: '' as Ref<Mixin<ScriptTypedPropertyEditorMixin<any>>>
+    VacancyList: '' as Ref<Mixin<VacancyList>>
   },
   attribute: {
     State: '' as Ref<Attribute<Status>>
